New article and journal abbreviations
Article fields that aren't recognized when including the references in a
document will be ignored.  So, I changed the full journal name field
from journal to fjournal and set the abbreviated journal names to the
journal field.  If I publish in a journal that requries the full journal
name I can just do a text search and replace to change fjournal back to
journal.